backend/requirements.txt
Django
djangorestframework
gunicorn
whitenoise
pillow
django-cors-headers